# docs/swagger.yaml
definitions:
  reggie_go_api_v1.CreateCategoryRequest:
    properties:
      name:
        description: 分类名称
        type: string
      sort:
        description: 顺序
        type: string
      type:
        description: 类型   1 菜品分类 2 套餐分类
        type: string
    type: object
  reggie_go_api_v1.CreateDishRequest:
    properties:
      categoryId:
        description: 菜品分类id
        type: string
      code:
        description: 商品码
        type: string
      description:
        description: 描述信息
        type: string
      flavors:
        description: 口味
        items:
          $ref: '#/definitions/reggie_go_api_v1.Flavors'
        type: array
      image:
        description: 图片
        type: string
      name:
        description: 菜品名称
        type: string
      price:
        description: 菜品价格
        type: number
      status:
        description: 0 停售 1 起售
        type: integer
    type: object
  reggie_go_api_v1.CreateEmployeeRequest:
    properties:
      idNumber:
        description: 身份证号
        type: string
      name:
        description: 姓名
        type: string
      phone:
        description: 手机号
        type: string
      sex:
        description: 性别
        type: string
      username:
        description: 用户名
        type: string
    type: object
  reggie_go_api_v1.DishFlavor:
    properties:
      createTime:
        description: 创建时间
        type: string
      createUser:
        description: 创建人
        type: integer
      dishId:
        description: 菜品
        type: integer
      id:
        description: 主键
        type: integer
      isDeleted:
        description: 是否删除
        type: integer
      name:
        description: 口味名称
        type: string
      updateTime:
        description: 更新时间
        type: string
      updateUser:
        description: 修改人
        type: integer
      value:
        description: 口味数据list
        type: string
    type: object
  reggie_go_api_v1.EmployeeLoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  reggie_go_api_v1.EmployeeLoginResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/reggie_go_api_v1.EmployeeLoginResponseData'
      map:
        additionalProperties: true
        type: object
      msg:
        type: string
    type: object
  reggie_go_api_v1.EmployeeLoginResponseData:
    properties:
      createTime:
        description: 创建时间
        type: string
      createUser:
        description: 创建人
        type: integer
      id:
        description: 主键
        type: integer
      idNumber:
        description: 身份证号
        type: string
      name:
        description: 姓名
        type: string
      password:
        description: 密码
        type: string
      phone:
        description: 手机号
        type: string
      sex:
        description: 性别
        type: string
      status:
        description: 状态 0:禁用，1:正常
        type: integer
      updateTime:
        description: 更新时间
        type: string
      updateUser:
        description: 修改人
        type: integer
      username:
        description: 用户名
        type: string
    type: object
  reggie_go_api_v1.Flavors:
    properties:
      name:
        description: 口味名称
        type: string
      showOption:
        type: boolean
      value:
        description: 口味数据list
        type: string
    type: object
  reggie_go_api_v1.GetCategoryPageData:
    properties:
      records:
        items:
          $ref: '#/definitions/reggie_go_internal_model.Category'
        type: array
      size:
        type: integer
      total:
        type: integer
    type: object
  reggie_go_api_v1.GetCategoryPageResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/reggie_go_api_v1.GetCategoryPageData'
      map:
        additionalProperties: true
        type: object
      msg:
        type: string
    type: object
  reggie_go_api_v1.GetDishByIdData:
    properties:
      categoryId:
        description: 菜品分类id
        type: integer
      code:
        description: 商品码
        type: string
      createTime:
        description: 创建时间
        type: string
      createUser:
        description: 创建人
        type: integer
      description:
        description: 描述信息
        type: string
      falavors:
        description: 口味
        items:
          $ref: '#/definitions/reggie_go_api_v1.DishFlavor'
        type: array
      id:
        description: 主键
        type: integer
      image:
        description: 图片
        type: string
      isDeleted:
        description: 是否删除
        type: integer
      name:
        description: 菜品名称
        type: string
      price:
        description: 菜品价格
        type: number
      sort:
        description: 顺序
        type: integer
      status:
        description: 0 停售 1 起售
        type: integer
      updateTime:
        description: 更新时间
        type: string
      updateUser:
        description: 修改人
        type: integer
    type: object
  reggie_go_api_v1.GetDishByIdResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/reggie_go_api_v1.GetDishByIdData'
      map:
        additionalProperties: true
        type: object
      msg:
        type: string
    type: object
  reggie_go_api_v1.GetDishByPageResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/reggie_go_api_v1.GetCategoryPageData'
      map:
        additionalProperties: true
        type: object
      msg:
        type: string
    type: object
  reggie_go_api_v1.GetEmployeeByIdData:
    properties:
      createTime:
        description: 创建时间
        type: string
      createUser:
        description: 创建人
        type: integer
      id:
        description: 主键
        type: integer
      idNumber:
        description: 身份证号
        type: string
      name:
        description: 姓名
        type: string
      password:
        description: 密码
        type: string
      phone:
        description: 手机号
        type: string
      sex:
        description: 性别
        type: string
      status:
        description: 状态 0:禁用，1:正常
        type: integer
      updateTime:
        description: 更新时间
        type: string
      updateUser:
        description: 修改人`
        type: integer
      username:
        description: 用户名
        type: string
    type: object
  reggie_go_api_v1.GetEmployeeByPaIdResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/reggie_go_api_v1.GetEmployeeByIdData'
      map:
        additionalProperties: true
        type: object
      msg:
        type: string
    type: object
  reggie_go_api_v1.GetEmployeeByPageData:
    properties:
      records:
        items:
          $ref: '#/definitions/reggie_go_internal_model.Employee'
        type: array
      size:
        type: integer
      total:
        type: integer
    type: object
  reggie_go_api_v1.GetEmployeeByPageResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/reggie_go_api_v1.GetEmployeeByPageData'
      map:
        additionalProperties: true
        type: object
      msg:
        type: string
    type: object
  reggie_go_api_v1.GetProfileResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/reggie_go_api_v1.GetProfileResponseData'
      map:
        additionalProperties: true
        type: object
      msg:
        type: string
    type: object
  reggie_go_api_v1.GetProfileResponseData:
    properties:
      nickname:
        example: alan
        type: string
      userId:
        type: string
    type: object
  reggie_go_api_v1.GetSetmealByPageResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/reggie_go_api_v1.GetEmployeeByPageData'
      map:
        additionalProperties: true
        type: object
      msg:
        type: string
    type: object
  reggie_go_api_v1.LoginRequest:
    properties:
      email:
        example: 1234@gmail.com
        type: string
      password:
        example: "123456"
        type: string
    required:
    - email
    - password
    type: object
  reggie_go_api_v1.LoginResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/reggie_go_api_v1.LoginResponseData'
      map:
        additionalProperties: true
        type: object
      msg:
        type: string
    type: object
  reggie_go_api_v1.LoginResponseData:
    properties:
      accessToken:
        type: string
    type: object
  reggie_go_api_v1.RegisterRequest:
    properties:
      email:
        example: 1234@gmail.com
        type: string
      password:
        example: "123456"
        type: string
    required:
    - email
    - password
    type: object
  reggie_go_api_v1.Response:
    properties:
      code:
        type: integer
      data: {}
      map:
        additionalProperties: true
        type: object
      msg:
        type: string
    type: object
  reggie_go_api_v1.UpdateCategoryRequest:
    properties:
      id:
        description: 主键
        type: string
      name:
        description: 分类名称
        type: string
      sort:
        description: 顺序
        type: integer
    type: object
  reggie_go_api_v1.UpdateEmployeeRequest:
    properties:
      createTime:
        description: 创建时间
        type: string
      createUser:
        description: 创建人
        type: integer
      id:
        description: 主键
        type: integer
      idNumber:
        description: 身份证号
        type: string
      name:
        description: 姓名
        type: string
      password:
        description: 密码
        type: string
      phone:
        description: 手机号
        type: string
      sex:
        description: 性别
        type: string
      status:
        description: 状态 0:禁用，1:正常
        type: integer
      updateTime:
        description: 更新时间
        type: string
      updateUser:
        description: 修改人
        type: integer
      username:
        description: 用户名
        type: string
    type: object
  reggie_go_internal_model.Category:
    properties:
      createTime:
        description: 创建时间
        type: string
      createUser:
        description: 创建人
        type: integer
      id:
        description: 主键
        type: integer
      name:
        description: 分类名称
        type: string
      sort:
        description: 顺序
        type: integer
      type:
        description: 类型   1 菜品分类 2 套餐分类
        type: integer
      updateTime:
        description: 更新时间
        type: string
      updateUser:
        description: 修改人
        type: integer
    type: object
  reggie_go_internal_model.Employee:
    properties:
      createTime:
        description: 创建时间
        type: string
      createUser:
        description: 创建人
        type: integer
      id:
        description: 主键
        type: integer
      idNumber:
        description: 身份证号
        type: string
      name:
        description: 姓名
        type: string
      password:
        description: 密码
        type: string
      phone:
        description: 手机号
        type: string
      sex:
        description: 性别
        type: string
      status:
        description: 状态 0:禁用，1:正常
        type: integer
      updateTime:
        description: 更新时间
        type: string
      updateUser:
        description: 修改人
        type: integer
      username:
        description: 用户名
        type: string
    type: object
host: localhost:8000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Nunu Example API
  version: 1.0.0
paths:
  /category:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 分类ID
        in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reggie_go_api_v1.Response'
      summary: 删除分类
      tags:
      - 分类模块
    post:
      consumes:
      - application/json
      parameters:
      - description: params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/reggie_go_api_v1.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reggie_go_api_v1.Response'
      summary: 新增分类
      tags:
      - 分类模块
    put:
      consumes:
      - application/json
      parameters:
      - description: params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/reggie_go_api_v1.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reggie_go_api_v1.Response'
      summary: 更新分类
      tags:
      - 分类模块
  /category/page:
    get:
      consumes:
      - application/json
      parameters:
      - description: 页数
        in: query
        name: page
        type: string
      - description: 每页数
        in: query
        name: size
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reggie_go_api_v1.GetCategoryPageResponse'
      summary: 分页查询
      tags:
      - 分类模块
  /dish:
    post:
      consumes:
      - application/json
      parameters:
      - description: params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/reggie_go_api_v1.CreateDishRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reggie_go_api_v1.Response'
      summary: 新增菜品
      tags:
      - 菜品管理
  /dish/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: 菜品Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reggie_go_api_v1.GetDishByIdResponse'
      summary: 通过id获取菜品信息
      tags:
      - 菜品管理
  /dish/page:
    get:
      consumes:
      - application/json
      parameters:
      - description: 页数
        in: query
        name: page
        type: string
      - description: 每页数
        in: query
        name: size
        type: string
      - description: 菜品名称
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reggie_go_api_v1.GetDishByPageResponse'
      summary: 分页查询
      tags:
      - 菜品管理
  /employee:
    post:
      consumes:
      - application/json
      parameters:
      - description: params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/reggie_go_api_v1.CreateEmployeeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reggie_go_api_v1.Response'
      summary: 新增员工
      tags:
      - 员工模块
    put:
      consumes:
      - application/json
      parameters:
      - description: params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/reggie_go_api_v1.UpdateEmployeeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reggie_go_api_v1.Response'
      summary: 更新员工
      tags:
      - 员工模块
  /employee/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: 员工id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reggie_go_api_v1.GetEmployeeByPaIdResponse'
      summary: 通过Id获取员工信息
      tags:
      - 员工模块
  /employee/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/reggie_go_api_v1.EmployeeLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reggie_go_api_v1.EmployeeLoginResponse'
      summary: 员工登录
      tags:
      - 员工模块
  /employee/logout:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reggie_go_api_v1.Response'
      summary: 员工退出
      tags:
      - 员工模块
  /employee/page:
    get:
      consumes:
      - application/json
      parameters:
      - description: 页数
        in: query
        name: page
        type: string
      - description: 每页数
        in: query
        name: size
        type: string
      - description: 员工姓名
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reggie_go_api_v1.GetEmployeeByPageResponse'
      summary: 分页查询
      tags:
      - 员工模块
  /login:
    post:
      consumes:
      - application/json
      parameters:
      - description: params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/reggie_go_api_v1.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reggie_go_api_v1.LoginResponse'
      summary: 账号登录
      tags:
      - 用户模块
  /register:
    post:
      consumes:
      - application/json
      description: 目前只支持邮箱登录
      parameters:
      - description: params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/reggie_go_api_v1.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reggie_go_api_v1.Response'
      summary: 用户注册
      tags:
      - 用户模块
  /setmeal/page:
    get:
      consumes:
      - application/json
      parameters:
      - description: 页数
        in: query
        name: page
        type: string
      - description: 每页数
        in: query
        name: size
        type: string
      - description: 套餐名称
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reggie_go_api_v1.GetSetmealByPageResponse'
      summary: 分页查询
      tags:
      - 套餐模块
  /user:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reggie_go_api_v1.GetProfileResponse'
      security:
      - Bearer: []
      summary: 获取用户信息
      tags:
      - 用户模块
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"